--- a/ruv-swarm/npm/audit-after.json
+++ b/ruv-swarm/npm/audit-after.json
@@ -1,0 +1,22 @@
+{
+  "auditReportVersion": 2,
+  "vulnerabilities": {},
+  "metadata": {
+    "vulnerabilities": {
+      "info": 0,
+      "low": 0,
+      "moderate": 0,
+      "high": 0,
+      "critical": 0,
+      "total": 0
+    },
+    "dependencies": {
+      "prod": 41,
+      "dev": 954,
+      "optional": 133,
+      "peer": 0,
+      "peerOptional": 0,
+      "total": 994
+    }
+  }
+}