--- a/ruv-swarm/npm/wasm/package.json
+++ b/ruv-swarm/npm/wasm/package.json
@@ -1,3 +1,18 @@
 {
-  "type": "module"
+  "name": "ruv-swarm-wasm",
+  "type": "module",
+  "collaborators": [
+    "rUv Contributors"
+  ],
+  "description": "WebAssembly bindings for ruv-swarm neural network orchestration",
+  "version": "1.0.5",
+  "license": "MIT OR Apache-2.0",
+  "files": [
+    "ruv_swarm_wasm_bg.wasm",
+    "ruv_swarm_wasm.js"
+  ],
+  "main": "ruv_swarm_wasm.js",
+  "sideEffects": [
+    "./snippets/*"
+  ]
 }